Refactor medsam trainer to simple sam trainer
