+++ micro_sam/training/__init__.py
@@ -4,5 +4,5 @@
 from .sam_trainer import SamTrainer, SamLogger
 from .util import ConvertToSamInputs, get_trainable_sam_model, identity
 from .joint_sam_trainer import JointSamTrainer, JointSamLogger
-from .medsam_trainer import MedSAMTrainer
+from .simple_sam_trainer import SimpleSamTrainer
 from .training import train_sam, train_sam_for_configuration, default_sam_loader, default_sam_dataset, CONFIGURATIONS
